Added a subscribe on the :failed event of @subscriber
At the moment, a failure to (re)connect to redis is not correctly
exposed by this engine. Failure occurs silently, without any messages.
By subscribing to the :failed event the PubSub-client, the Faye-Redis
engine is notified when a connection failure is happening, and will
raise an error, which is useful for users.

This commit came to life after a hunt for a non-functioning deployment
in which ultimately a non-started Redis database was to blame. It would
have been nice if an error would have been raised.